src/backend/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
gunicorn==21.2.0
redis==5.0.1
PyJWT==2.8.0
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.2.0
